requirements.txt
einops==0.8.0
matplotlib==3.8.2
numpy==2.2.1
Pillow==11.1.0
torch==2.1.2
torchvision==0.11.3
tqdm==4.66.4
